requirements.txt
discord.py
gspread
fuzzywuzzy